server/requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
decorator==4.1.1
Flask==0.12.2
Flask-HTTPAuth==3.2.3
Flask-JWT==0.3.2
Flask-Script==2.0.5
Flask-SocketIO==2.9.1
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
gevent==1.2.2
greenlet==0.4.12
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.10
passlib==1.7.1
pbr==3.1.1
pkg-resources==0.0.0
PyJWT==1.4.2
PyMySQL==0.7.11
python-engineio==1.7.0
python-slugify==1.2.4
python-socketio==1.7.6
requests==2.18.1
scrypt==0.8.0
six==1.10.0
socketIO-client==0.7.2
SQLAlchemy==1.1.11
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Unidecode==0.4.21
urllib3==1.21.1
websocket-client==0.44.0
Werkzeug==0.12.2
